agent_workspace/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic>=2.5.2,<3
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
aiohttp==3.9.5
xmltodict==0.13.0
